Rename Homepage classes.
