+++ www/docs/scotland/index.php
@@ -2,7 +2,7 @@
 
 include_once '../../includes/easyparliament/init.php';
 
-$view = new MySociety\TheyWorkForYou\SPHomepage();
+$view = new MySociety\TheyWorkForYou\Homepage\Scotland();
 $data = $view->display();
 
 MySociety\TheyWorkForYou\Renderer::output('scotland/index', $data);
